Move from extending Observable to StreamTransformers
