+++ lib/rxdart.dart
@@ -1,10 +1,20 @@
 library rx;
 
 export 'package:rxdart/src/observable.dart'
-    show Observable, observable, Ease, TimeInterval, GroupByMap;
-
-export 'package:rxdart/src/operators/of_type.dart'
-    show TypeToken, kBool, kDouble, kInt, kNum, kString, kSymbol;
+    show
+        Observable,
+        observable,
+        Ease,
+        TimeInterval,
+        GroupByMap,
+        TypeToken,
+        RetryError,
+        kBool,
+        kDouble,
+        kInt,
+        kNum,
+        kString,
+        kSymbol;
 
 export 'package:rxdart/src/subject/behaviour_subject.dart'
     show BehaviourSubject;
