--- lib/src/operators/with_latest_from.dart
@@ -1,52 +1,0 @@
-import 'package:rxdart/src/observable.dart';
-
-class WithLatestFromObservable<T, S, R> extends Observable<R> {
-  WithLatestFromObservable(
-      Stream<T> stream, Stream<S> latestFromStream, R fn(T t, S s))
-      : super(buildStream(stream, latestFromStream, fn));
-
-  static Stream<R> buildStream<T, S, R>(
-      Stream<T> stream, Stream<S> latestFromStream, R fn(T t, S s)) {
-    Stream<R> transformed = stream.transform(
-        new StreamTransformer<T, R>((Stream<T> input, bool cancelOnError) {
-      StreamController<R> controller;
-      StreamSubscription<T> subscription;
-      StreamSubscription<S> latestFromSubscription;
-      S latestValue;
-
-      controller = new StreamController<R>(
-          sync: true,
-          onListen: () {
-            subscription = input.listen((T value) {
-              if (latestValue != null) {
-                controller.add(fn(value, latestValue));
-              }
-            }, onError: controller.addError);
-
-            latestFromSubscription = latestFromStream.listen((S latest) {
-              latestValue = latest;
-            },
-                onError: controller.addError,
-                onDone: controller.close,
-                cancelOnError: cancelOnError);
-          },
-          onPause: ([Future<dynamic> resumeSignal]) =>
-              subscription.pause(resumeSignal),
-          onResume: () => subscription.resume(),
-          onCancel: () {
-            return Future.wait(<Future<dynamic>>[
-              subscription.cancel(),
-              latestFromSubscription.cancel()
-            ].where((Future<dynamic> cancelFuture) => cancelFuture != null));
-          });
-
-      return controller.stream.listen(null);
-    }));
-
-    if (stream.isBroadcast) {
-      return transformed.asBroadcastStream();
-    } else {
-      return transformed;
-    }
-  }
-}
