+++ lib/src/subject/behaviour_subject.dart
@@ -1,5 +1,3 @@
-import 'dart:async';
-
 import 'package:rxdart/src/observable.dart';
 
 class _BehaviourSink<T> implements EventSink<T> {
