+++ lib/src/subject/replay_subject.dart
@@ -1,4 +1,3 @@
-import 'dart:async';
 import 'dart:collection';
 
 import 'package:rxdart/src/observable.dart';
